requirements.txt
annotated-types>=0.6.0
anyio>=4.2.0
backports.tarfile>=1.2.0
bcrypt>=4.1.2
bottle>=0.12.25
certifi>=2024.8.30
cffi>=1.17.1
charset-normalizer>=3.4.0
click>=8.1.7
clr-loader>=0.2.6
colorama>=0.4.6
contourpy>=1.3.0
cryptography>=41.0.7
cycler>=0.12.1
et_xmlfile>=2.0.0
exceptiongroup>=1.2.0
fastapi>=0.109.0
Flask>=2.2.2
fonttools>=4.55.2
future>=1.0.0
h11>=0.14.0
httptools>=0.6.1
idna>=3.6
igraph>=0.10.8
importlib_metadata>=8.5.0
importlib_resources>=6.4.5
itsdangerous>=2.2.0
jaraco.classes>=3.4.0
jaraco.context>=6.0.1
jaraco.functools>=4.1.0
Jinja2>=3.1.3
junos-eznc>=2.7.2
keyring>=25.5.0
kiwisolver>=1.4.7
ldap3>=2.9.1
lxml>=5.3.0
MarkupSafe>=2.1.3
matplotlib>=3.9.3
more-itertools>=10.5.0
n2g>=0.3.3
napalm>=5.0.0
ncclient>=0.6.15
netaddr>=1.3.0
netmiko>=4.4.0
netutils>=1.10.0
networkx>=3.2.1
ntc_templates>=7.4.0
numpy>=2.0.2
openpyxl>=3.1.5
packaging>=24.1
paramiko>=3.5.0
passlib>=1.7.4
pillow>=11.0.0
proxy_tools>=0.1.0
pyasn1>=0.6.1
pycparser>=2.21
pydantic>=2.5.3
pydantic_core>=2.14.6
pyeapi>=1.0.4
PyJWT>=2.9.0
PyNaCl>=1.5.0
pynetbox>=7.4.0
pynxos>=0.0.5
pyparsing>=3.2.0
PyQt6>=6.7.1
PyQt6-Charts>=6.7.0
PyQt6-Charts-Qt6>=6.7.3
PyQt6-Qt6>=6.7.3
PyQt6-WebEngine>=6.7.0
PyQt6-WebEngine-Qt6>=6.7.3
PyQt6-WebEngineSubwheel-Qt6>=6.7.3
PyQt6_sip>=13.8.0
pyserial>=3.5
python-dateutil>=2.9.0.post0
python-dotenv>=1.0.0
python-igraph>=0.10.8
python-multipart>=0.0.6
pythonnet>=3.0.3
pywebview>=4.4.1
pywin32-ctypes>=0.2.3
PyYAML>=6.0.1
requests>=2.32.3
scp>=0.15.0
six>=1.17.0
sniffio>=1.3.0
starlette>=0.35.1
svgwrite>=1.4.3
textfsm>=1.1.3
texttable>=1.7.0
transitions>=0.9.2
ttp>=0.9.5
ttp-templates>=0.3.7
typing_extensions>=4.9.0
urllib3>=2.2.3
uvicorn>=0.26.0
watchfiles>=0.21.0
websockets>=12.0
Werkzeug>=3.1.3
yamlordereddictloader>=0.4.2
zipp>=3.21.0
